Phone number called extracted to config files.
In SDS.test.test_vio_vad_jasr_ftts, the extension of the phone number
called to was hard-wired, and so was the rest of the URI. I made it be
read from configuration, and by the way reniced the way configuration is
handled. Especially, imports needed for a given configuration do not
clog the SDS.utils.config module, but rather they have to be present in
the corresponding config file that requires them.