Added capitalization markings to templates
If '/Cap1' is added after the slot name in the template, the first letter
of the value is capitalized on the output.

This is the only reliable way to solve capitalization in templates.
The output may consist of multiple templates and/or multiple sentences.
However, the templates usually either begin or do not begin a sentence
in any case.

The values may be inflected and modified in many different ways using
this mechanism in future.