Fix for CRWS not returning the _sNum2 attribute
CRWS may not indicate the _sNum2 attribute in some TrainInfo instances.
The SUDS library will not initialize it to any default value and keep it
as missing. This checks the presence of this attribute before it is
used.

Additional small bugfixes to debug logging in DummyLogger, comments only
in SessionLogger.