Fixing HDC-policy to avoid "Zastávku XY v městě none neznám."
If there are more possible cities both for from_stop and for to_stop,
the policy passed from_city == 'none' and to_city == 'none' on without
checking, which lead to confusing prompts, such as:

> Promiňte, asi Vám nerozumím. Zastávku Budějovická v obci none neznám.
("I am sorry, but I don't understand. I know of no stop called
Budějovická in the city of none")

This commit fixes the problem.