Fixing weird behavior when the user requests something straight away
If the user wanted a specific information about the route right at the
beginning of the dialogue, Alex would say that it doesn't know the
origin and destination stops, even though the user provided this in the
sentence.

This fix first tries to search for a connection, then returns the
specified info (or an error if the connection search failed because of
missing information).