Inflecting city names if found in dictionary
Inflecting city names if they are found in the Morphodita dictionary,
i.e., instead of ``z obce Praha``, the NLG component can now output
``z Prahy``. It falls back to ``z obce XY`` if the city name is not in
the dictionary and the generator does not know how to inflect it.

This can be easily extended to stop names, but I'm not sure if we
want that.