route_alternative moved from slots to a special state variable
The `dialogue_state` variable is now a member variable of the state,
(same as `conn_info` and `directions`), not a slot.

This reflects the actual usage of the variable – it is always integer,
never a probability distribution. Now wo don't need to use `isinstance`
to test whether a route has been found, we can use `is not None`, which
is nicer.

This fixes the problem reported in #165.