Fix for #168 – avoiding the city of 'none'
This fixes the bug reported in #168. It had probably been introduced
in fbea940 -- `stop_city_inferred` is true if any of the stops has
a city inferred, which may not hold for the other stop, so it's not
safe to use it at this place.

This also means that now we always ask for the stop first and only
when we're unsure (if the stop is in several different cities), we ask
for the city.