Improved code handling
Now the client is prepared for the situation that some code may already
be in use and checks the server response on adding the key.
If a key is in use, 10 attempts are made at generating a different key.
If no valid key can be generated, an apology is issued.