Remembering call log directory and task in CF CSV
When the CF job asks for a task, it now not only shows the task to the
user, but also stores it in a hidden input field.

The call log directory is now passed from Alex to the server when
storing a code. When the code is validated, the directory is passed on
to the CF job (and stored in a hidden input field).